Move check for illegal base into derivation model
Ref. #456
